requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
environ==1.0
setuptools==69.1.1
sqlparse==0.5.0
typing_extensions==4.12.2
wheel==0.43.0
psycopg2-binary==2.9.9
djangorestframework==3.15.1